+++ src/suou/exceptions.py
@@ -14,8 +14,6 @@ This software is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 """
 
-from .functools import deprecated
-
 class MissingConfigError(LookupError):
     """
     Config variable not found.
@@ -42,3 +40,7 @@ class InconsistencyError(RuntimeError):
     """
     This program is in a state which it's not supposed to be in.
     """
+
+__all__ = (
+    'MissingConfigError', 'MissingConfigWarning', 'LexError', 'InconsistencyError'
+)